+++ android/src/main/kotlin/io/vinicius/appodeal_flutter/AppodealFlutterPlugin.kt
@@ -3,6 +3,10 @@ package io.vinicius.appodeal_flutter
 import android.app.Activity
 import androidx.annotation.NonNull
 import com.appodeal.ads.Appodeal
+import com.appodeal.ads.BannerCallbacks
+import com.appodeal.ads.InterstitialCallbacks
+import com.appodeal.ads.NonSkippableVideoCallbacks
+import com.appodeal.ads.RewardedVideoCallbacks
 import com.explorestack.consent.Consent
 import com.explorestack.consent.ConsentForm
 import com.explorestack.consent.ConsentFormListener
@@ -67,6 +71,9 @@ class AppodealFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware
         val adTypes = args["adTypes"] as List<Int>
         val testMode = args["testMode"] as Boolean
 
+        // Registering callbacks
+        setCallbacks()
+
         val ads = adTypes.fold(0) { acc, value -> acc or getAdType(value) }
         Appodeal.setTesting(testMode)
         Appodeal.initialize(activity, appKey, ads, hasConsent)
@@ -89,16 +96,127 @@ class AppodealFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware
     }
     // endregion
 
-    // region - Helper
-    private fun getAdType(adId: Int): Int {
-        return when (adId) {
-            1 -> Appodeal.BANNER
-            2 -> Appodeal.NATIVE
-            3 -> Appodeal.INTERSTITIAL
-            4 -> Appodeal.REWARDED_VIDEO
-            5 -> Appodeal.NON_SKIPPABLE_VIDEO
-            else -> Appodeal.NONE
-        }
+    // region - Callbacks
+    private fun setCallbacks() {
+        Appodeal.setBannerCallbacks(object : BannerCallbacks {
+            override fun onBannerLoaded(p0: Int, p1: Boolean) {
+                channel.invokeMethod("onBannerLoaded", null)
+            }
+
+            override fun onBannerFailedToLoad() {
+                channel.invokeMethod("onBannerFailedToLoad", null)
+            }
+
+            override fun onBannerShown() {
+                channel.invokeMethod("onBannerShown", null)
+            }
+
+            override fun onBannerShowFailed() {
+                // Not implemented for the same o consistency with iOS
+            }
+
+            override fun onBannerClicked() {
+                channel.invokeMethod("onBannerClicked", null)
+            }
+
+            override fun onBannerExpired() {
+                channel.invokeMethod("onBannerExpired", null)
+            }
+        })
+
+        Appodeal.setInterstitialCallbacks(object : InterstitialCallbacks {
+            override fun onInterstitialLoaded(isPrecache: Boolean) {
+                channel.invokeMethod("onInterstitialLoaded", null)
+            }
+
+            override fun onInterstitialFailedToLoad() {
+                channel.invokeMethod("onInterstitialFailedToLoad", null)
+            }
+
+            override fun onInterstitialShown() {
+                channel.invokeMethod("onInterstitialShown", null)
+            }
+
+            override fun onInterstitialShowFailed() {
+                channel.invokeMethod("onInterstitialShowFailed", null)
+            }
+
+            override fun onInterstitialClicked() {
+                channel.invokeMethod("onInterstitialClicked", null)
+            }
+
+            override fun onInterstitialClosed() {
+                channel.invokeMethod("onInterstitialClosed", null)
+            }
+
+            override fun onInterstitialExpired() {
+                channel.invokeMethod("onInterstitialExpired", null)
+            }
+        })
+
+        Appodeal.setRewardedVideoCallbacks(object : RewardedVideoCallbacks {
+            override fun onRewardedVideoLoaded(isPrecache: Boolean) {
+                channel.invokeMethod("onRewardedVideoLoaded", null)
+            }
+
+            override fun onRewardedVideoFailedToLoad() {
+                channel.invokeMethod("onRewardedVideoFailedToLoad", null)
+            }
+
+            override fun onRewardedVideoShown() {
+                channel.invokeMethod("onRewardedVideoShown", null)
+            }
+
+            override fun onRewardedVideoShowFailed() {
+                channel.invokeMethod("onRewardedVideoShowFailed", null)
+            }
+
+            override fun onRewardedVideoFinished(p0: Double, p1: String?) {
+                channel.invokeMethod("onRewardedVideoFinished", null)
+            }
+
+            override fun onRewardedVideoClosed(p0: Boolean) {
+                channel.invokeMethod("onRewardedVideoClosed", null)
+            }
+
+            override fun onRewardedVideoExpired() {
+                channel.invokeMethod("onRewardedVideoExpired", null)
+            }
+
+            override fun onRewardedVideoClicked() {
+                channel.invokeMethod("onRewardedVideoClicked", null)
+            }
+        })
+
+        Appodeal.setNonSkippableVideoCallbacks(object : NonSkippableVideoCallbacks {
+            override fun onNonSkippableVideoLoaded(p0: Boolean) {
+                channel.invokeMethod("onNonSkippableVideoLoaded", null)
+            }
+
+            override fun onNonSkippableVideoFailedToLoad() {
+                channel.invokeMethod("onNonSkippableVideoFailedToLoad", null)
+            }
+
+            override fun onNonSkippableVideoShown() {
+                channel.invokeMethod("onNonSkippableVideoShown", null)
+            }
+
+            override fun onNonSkippableVideoShowFailed() {
+                channel.invokeMethod("onNonSkippableVideoShowFailed", null)
+            }
+
+            override fun onNonSkippableVideoFinished() {
+                channel.invokeMethod("onNonSkippableVideoFinished", null)
+            }
+
+            override fun onNonSkippableVideoClosed(p0: Boolean) {
+                channel.invokeMethod("onNonSkippableVideoClosed", null)
+            }
+
+            override fun onNonSkippableVideoExpired() {
+                channel.invokeMethod("onNonSkippableVideoExpired", null)
+            }
+        })
     }
     // endregion
 
@@ -149,6 +267,19 @@ class AppodealFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware
                 .build()
 
         consentForm.load()
+    }
+    // endregion
+
+    // region - Helper Methods
+    private fun getAdType(adId: Int): Int {
+        return when (adId) {
+            1 -> Appodeal.BANNER
+            2 -> Appodeal.NATIVE
+            3 -> Appodeal.INTERSTITIAL
+            4 -> Appodeal.REWARDED_VIDEO
+            5 -> Appodeal.NON_SKIPPABLE_VIDEO
+            else -> Appodeal.NONE
+        }
     }
     // endregion
 }